refactor: isolate the path-scoped push check
